+++ plugins/terminal/src/lib.rs
@@ -1,8 +1,3 @@
 mod plugin;
 
-pub use plugin::*;
-
-#[no_mangle]
-pub extern "C" fn create_plugin() -> Box<dyn playground_core_plugin::Plugin> {
-    Box::new(plugin::create())
-}
+pub use plugin::TerminalPlugin;